Entrata Leases

This function retrieves lease information from the Entrata API.

Usage

entrata_leases(
  property_id,
  application_id = as.integer(NULL),
  customer_id = as.integer(NULL),
  lease_status_type_ids = as.integer(c(NULL)),
  lease_ids = as.integer(c(NULL)),
  scheduled_ar_code_ids = as.integer(c(NULL)),
  unit_number = as.character(NULL),
  building_name = as.character(NULL),
  move_in_date_from = as.Date(NULL),
  move_in_date_to = as.Date(NULL),
  lease_expiring_date_from = as.Date(NULL),
  lease_expiring_date_to = as.Date(NULL),
  move_out_date_from = as.Date(NULL),
  move_out_date_to = as.Date(NULL),
  include_other_income_leases = FALSE,
  resident_friendly_mode = FALSE,
  include_lease_history = FALSE,
  include_ar_transactions = FALSE,
  pagination_page_number = 1,
  pagination_page_size = 500,
  include_pagination_links = FALSE,
  ...
)

Arguments

application_id

Optional. Integer value of the application ID

customer_id

Optional. Integer value of the customer ID

lease_status_type_ids

Optional. Integer vector of lease status type IDs

lease_ids

Optional. Integer vector of lease IDs

scheduled_ar_code_ids

Optional. Integer vector of scheduled AR code IDs

unit_number

Optional. Character value of the unit number

building_name

Optional. Character value of the building name

move_in_date_from

Optional. Date value of the move-in date from

move_in_date_to

Optional. Date value of the move-in date to

lease_expiring_date_from

Optional. Date value of the lease expiring from

lease_expiring_date_to

Optional. Date value of the lease expiring to

move_out_date_from

Optional. Date value of the move-out date from

move_out_date_to

Optional. Date value of the move-out date to

include_other_income_leases

Optional. Logical value to include other income leases

resident_friendly_mode

Optional. Logical value to include resident friendly mode

include_lease_history

Optional. Logical value to include lease history information

include_ar_transactions

Optional. Logical value to include AR transactions

pagination_page_number

Pagination page number. Default is 1

pagination_page_size

Number of items per page. Default is 500

include_pagination_links

Logical value to include pagination links in the response. Default is FALSE

...

Additional parameters to pass to the request

property_ids

Required. Integer vector of property IDs

Value

Parsed Response Body Content as a tibble with leases data.

See also