Generate Pre-Lease Report

This function generates a pre-lease report in Entrata based on the provided parameters and returns the summary and details of the report.

Usage

entrata_pre_lease_report(
  property_ids = c(NULL),
  period_start = "09/01/2024",
  ...
)

Arguments

property_ids

A vector of property IDs to include in the report.

period_start

The start date of the reporting period. Defaults to "09/01/2024".

...

Additional parameters for the report.

Value

A list containing the summary and details of the pre-lease report.