Entrata API Properties Endpoint

This function retrieves properties from the Entrata API.

This function parses the response from the Entrata API's "getProperties" method.

Usage

entrata_properties(
  property_ids = c(NULL),
  property_lookup_codes = NULL,
  show_all_status = FALSE,
  ...
)

parse_property_res_data(res_data)

Arguments

property_ids

Character Vector of Property IDs to include in the request body parameters. If NULL, all properties will be returned. Default is NULL.

property_lookup_codes

Character string with a "Property Lookup Code" to include in the request's body. Default is NULL.

show_all_status

Logical: if TRUE will return all properties, regardless of status. Default is FALSE.

...

Arguments passed on to entrata

endpoint

Entrata API Endpoint to send the request to. Default is NULL. See details for the available endpoints.

method

Entrata API Method to use in the request. Not to be confused with the HTTP request method (i.e. GET, POST), this method must be a method that is available in the Entrata API for the specified endpoint. The default value depends on whether the endpoint parameter was provided, if it is available the default value is the first method listed for that endpoint. If no endpoint is specified, NULL will be used as the method in the request body. See details for available methods by endpoint and more details on the structure of the request.

method_version

Entrata API Method Version to use in the request. Default is "r1". Some endpoints have multiple versions of the same method. The version should be provided as a string. See details for more information on the structure of the request and available versions by endpoint method.

method_params

List of parameters to use in the request body's "method" object. Default is an empty list. The parameters should be provided as a named list where the names are the parameter names and the values are the parameter values. See details for more information on the structure of the request and available parameters by endpoint method. The default value depends on whether the endpoint parameter was provided, if it is available the default value is an empty list. If no endpoint is specified, NULL will be used as the method in the request body.

ua

User Agent string to use in the request. Default is to use user_agent().

perform

Logical value indicating whether to perform the request. Default is FALSE. If FALSE, the function will return the request object without performing the request. If TRUE, the function will perform the request and return the response object.

extract

Logical value indicating whether to extract the response. Default is the value of perform. If TRUE, the function will extract the response object and return it. If FALSE, the function will return the response object as is.

enable_retry

Logical value indicating whether to enable request retry. Default is FALSE. If TRUE, the function will enable request retry with the default retry settings. If FALSE, the function will not enable request retry.

timeout

Numeric value indicating the request timeout in seconds. Default is NULL. If provided, the function will set the request timeout to the provided value in seconds.

dry_run

Logical value indicating whether to perform a dry run of the request. Default is FALSE. If TRUE, the function will perform a dry run of the request before performing the actual request (or if perform is not set, will return the request object without performing the request).

progress

Logical value indicating whether to show progress of the request. Only useful for long running requests. Default is FALSE. If TRUE, the function will show the progress of the request.

config

Entrata API Configuration Values as a list. Default is to use config::get("entrata") to retrieve the configuration values from a config.yml configuration file. The configuration values should include the following keys: username, password, and base_url. See details.

verbosity

How much information to print? This is a wrapper around req_verbose() that uses an integer to control verbosity:

0: no output

1: show headers

2: show headers and bodies

3: show headers, bodies, and curl status messages.

Use with_verbosity() to control the verbosity of requests that you can't affect directly.

res_data

Response data from the Entrata API's "getProperties" method.

Value

A tibble with the property information

A tibble with the property information

Functions

parse_property_res_data(): Parse Entrata Properties Response Data

See also

Examples