Prepare Pre-Lease Report Parameters

This function prepares the parameters required for generating a pre-lease report in Entrata.

Usage

prep_pre_lease_report_params(
  latest_report_version = mem_get_latest_report_version("pre_lease"),
  property_group_ids = mem_get_property_ids_filter_param(),
  period_start_date = "09/01/2024",
  period_type = "date",
  summarize_by = "unit_type",
  group_by = "unit_type",
  consider_pre_leased_on = "33",
  charge_code_detail = 1,
  space_options = "do_not_show",
  additional_units_shown = "available",
  combine_unit_spaces_with_same_lease = 0,
  consolidate_by = "no_consolidation",
  arrange_by_property = 0,
  subtotals = list("summary", "details"),
  yoy = 1,
  ...
)

Arguments

latest_report_version

A character string representing the latest version of the pre-lease report. Defaults to fetching the latest version.

property_group_ids

A list of property group IDs to include in the report.

period_start_date

The start date of the reporting period. Defaults to "09/01/2024".

period_type

The type of period for the report. Defaults to "date".

summarize_by

The level at which to summarize the report. Defaults to "unit_type".

group_by

The level at which to group the report. Defaults to "unit_type".

consider_pre_leased_on

A string representing the pre-leased on date. Defaults to "33".

charge_code_detail

An integer for charge code details. Defaults to 1.

space_options

A string representing space options. Defaults to "do_not_show".

additional_units_shown

A string representing additional units shown. Defaults to "available".

combine_unit_spaces_with_same_lease

An integer indicating whether to combine unit spaces. Defaults to 0.

consolidate_by

A string indicating consolidation options. Defaults to "no_consolidation".

arrange_by_property

An integer indicating whether to arrange by property. Defaults to 0.

subtotals

A list of subtotals to include in the report. Defaults to c("summary", "details").

yoy

An integer indicating whether to include year-over-year data. Defaults to 1.

...

Additional parameters for the report.

Value

A list of parameters for the pre-lease report.